Refactored board creation
Ok…

Just for testing purposes, the board manager uses now only a group of 4
chunks disposed in a square-like formation.

I made this in a way that our Show-Only-Visible-Chunks function still
works as it was working for single chunks.

Also fixed a bug in this function that happened only on non-square map
sizes.